refactor(events): extract contact_email helper to dedupe across checkout routes
