+++ apps/events/app/api/checkout/route.ts
@@ -15,6 +15,7 @@ import { eq, and, sql } from 'drizzle-orm';
 import { optionalAuth } from '@imajin/auth';
 import { rateLimit, getClientIP } from '@/src/lib/rate-limit';
 import { getClient } from '@imajin/db';
+import { getContactEmail } from '@/src/lib/contact-email';
 
 const PAY_SERVICE_URL = process.env.PAY_SERVICE_URL!;
 const EVENTS_URL = process.env.NEXT_PUBLIC_EVENTS_URL!;
@@ -116,15 +117,8 @@ export const POST = withLogger('events', async (request, { log, correlationId })
     // Resolve contact email for pre-filling Stripe checkout
     let customerEmail = body.email;
     if (buyerDid && !customerEmail) {
-      try {
-        const sql = getClient();
-        const rows = await sql<{ contact_email: string | null }[]>`
-          SELECT contact_email FROM auth.identities WHERE id = ${buyerDid} LIMIT 1
-        `;
-        customerEmail = rows[0]?.contact_email ?? undefined;
-      } catch (err) {
-        // non-fatal — Stripe will collect email on its page
-      }
+      const resolved = await getContactEmail(buyerDid, log);
+      if (resolved) customerEmail = resolved;
     }
 
     // Enforce max per order: per-type override > event metadata > default 10, hard cap 20
